Adapt rest of the address structure section
With adding multi-sig addresses, we'll need to update how HD wallets work. Made
a todo item for that. There is a way to get HD wallets to work with the
script-based multi-sig addresses, using a slight variation of BIP-45. I've
discussed that with @dcoutts, but still need to write down the particulars.